\name{BiplotGUI-package}
\alias{BiplotGUI-package}
\alias{BiplotGUI}
\docType{package}
\title{
BiplotGUI: Interactive Biplots in R
}
\description{
Provides a GUI with which users can construct and interact with biplots.
}
\details{
\tabular{ll}{
Package: \tab BiplotGUI\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2008-08-12\cr
License: \tab GPL-3 or later\cr
LazyLoad: \tab yes\cr
}
The \code{Biplots} function initialises the GUI.

At present, BiplotGUI is written to run under Windows. The GUI runs better when R is in SDI mode rather than MDI mode.
}
\author{
Author and maintainer: Anthony la Grange <amlg at sun.ac.za>, <\url{http://biplotgui.r-forge.r-project.org/}>
}
\references{
  Gower JC, Hand DJ (1996). \emph{Biplots}. Monographs on Statistics and Applied Probability. Chapman \& Hall, London, UK.
}
\keyword{ package }
