\name{BiplotGUI-package}
\alias{BiplotGUI-package}
\alias{BiplotGUI}
\docType{package}
\title{
BiplotGUI: Interactive Biplots in R
}
\description{
A GUI with which to construct and interact with biplots.
}
\details{
\tabular{ll}{
Package: \tab BiplotGUI\cr
Type: \tab Package\cr
Version: \tab 0.0-6\cr
Date: \tab 2010-02-14\cr
License: \tab GPL-3 or later\cr
LazyLoad: \tab yes\cr
}
The \code{Biplots} function initialises the GUI.

At present, \pkg{BiplotGUI} is intended to be run under Windows. In Windows, it runs marginally better in SDI mode, rather than MDI mode. 
}
\author{
Author and maintainer: Anthony la Grange 

Includes Fortran code adapted by NJ le Roux from the original by PJ Rousseeuw, I Ruts and JW Tukey.
}
\references{
  Gower JC, Hand DJ (1996). \emph{Biplots}. Monographs on Statistics and Applied Probability. Chapman \& Hall, London, UK.
}
\keyword{ package }
